Factor out node_config
